Fixed bug in datepublisher page extension
The recently added granular_now datetime.now replacement caused a bug
where a newly created page would be inactive on publishing for as much
as 4 minutes 59 seconds. This is because datetime.now was used to set
the publication date, but this publication date was being compared
against the floor divided granular_now - which would always be smaller
on publication - when determining if that page should be active.